Add check for DNS lookup restriction
DNS lookups will not be initiated to the authoritative nameservers
for the domain if such requests are blocked by firewall. The check
takes `dig timeout` (i.e. 5 seconds) if the requests are blocked.

Also simplify alternates mechanism in mkblocks.sh to not query
authoritative nameservers so that the only script which stays
affected by firewalling DNS is `despf.sh`.